CORDA-2491 Add entry point with generic extra configuration options. Move configuration verification code to avoid circular dependencies.
