+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/network/NetworkParameterOverridesSpec.kt
@@ -1,4 +1,4 @@
-package net.corda.bootstrapper
+package net.corda.nodeapi.internal.network
 
 import com.typesafe.config.Config
 import net.corda.common.configuration.parsing.internal.Configuration
@@ -9,15 +9,13 @@ import net.corda.common.validation.internal.Validated
 import net.corda.core.internal.noPackageOverlap
 import net.corda.core.internal.requirePackageValid
 import net.corda.nodeapi.internal.crypto.loadKeyStore
-import net.corda.nodeapi.internal.network.NetworkParametersOverrides
-import net.corda.nodeapi.internal.network.PackageOwner
 import java.io.IOException
 import java.nio.file.InvalidPathException
 import java.nio.file.Path
 import java.nio.file.Paths
 import java.security.KeyStoreException
 
-internal typealias Valid<TARGET> = Validated<TARGET, Configuration.Validation.Error>
+typealias Valid<TARGET> = Validated<TARGET, Configuration.Validation.Error>
 
 fun Config.parseAsNetworkParametersConfiguration(options: Configuration.Validation.Options = Configuration.Validation.Options(strict = false)):
         Valid<NetworkParametersOverrides> = NetworkParameterOverridesSpec.parse(this, options)
@@ -26,15 +24,15 @@ internal fun <T> badValue(msg: String): Valid<T> = Validated.invalid(sequenceOf(
 internal fun <T> valid(value: T): Valid<T> = Validated.valid(value)
 
 internal object NetworkParameterOverridesSpec : Configuration.Specification<NetworkParametersOverrides>("DefaultNetworkParameters") {
-    private val minimumPlatformVersion by int().mapValid(::parsePositiveInteger).optional()
-    private val maxMessageSize by int().mapValid(::parsePositiveInteger).optional()
-    private val maxTransactionSize by int().mapValid(::parsePositiveInteger).optional()
+    private val minimumPlatformVersion by int().mapValid(NetworkParameterOverridesSpec::parsePositiveInteger).optional()
+    private val maxMessageSize by int().mapValid(NetworkParameterOverridesSpec::parsePositiveInteger).optional()
+    private val maxTransactionSize by int().mapValid(NetworkParameterOverridesSpec::parsePositiveInteger).optional()
     private val packageOwnership by nested(PackageOwnershipSpec).list().optional()
     private val eventHorizon by duration().optional()
 
     internal object PackageOwnershipSpec : Configuration.Specification<PackageOwner>("PackageOwners") {
-        private val packageName by string().mapValid(::toPackageName)
-        private val keystore by string().mapValid(::toPath)
+        private val packageName by string().mapValid(PackageOwnershipSpec::toPackageName)
+        private val keystore by string().mapValid(PackageOwnershipSpec::toPath)
         private val keystorePassword by string()
         private val keystoreAlias by string()
